Clever includes

+++ inc/raycaster.h
@@ -3,14 +3,9 @@
 
 #include <stdbool.h>
 
-#include "player.h"
-#include "map.h"
-#include "trigonometry.h"
-#include "render.h"
-
-#define FOV 60                                     /** Field of view (between 60 and 90 is okay, more than 90 warps the image)                */
-#define AMMOUNT_OF_RAYS ((FOV * 2) * resolution)   /** Ammount of rays per degree of FOV (more rays means more image quality, normal is 2)    */
-#define RATIO_ANGLE_RAYS (AMMOUNT_OF_RAYS / FOV)   /** Relation between AMMOUNT_OF_RAYS and FOV, for making the right steps between every ray */
+#define FOV 60                                             /** Field of view (between 60 and 90 is okay, more than 90 warps the image)                */
+#define AMMOUNT_OF_RAYS ((FOV * 2) * resolution)           /** Ammount of rays per degree of FOV (more rays means more image quality, normal is 2)    */
+#define RATIO_ANGLE_RAYS (AMMOUNT_OF_RAYS / (float) FOV)   /** Relation between AMMOUNT_OF_RAYS and FOV, for making the right steps between every ray */
 
 #define NORTH_SOUTH_WALL 0     /** The North / South orientation of the wall  */
 #define WEAST_EAST_WALL  1     /** The Weast / East orientation of the wall */
